% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_ttz.R
\name{classify_ttz}
\alias{classify_ttz}
\title{Identify Dispersal Vanguards and Laggards in the Temporary Transition Zone}
\usage{
classify_ttz(x)
}
\arguments{
\item{x}{The matrix of microbial abundance.}
}
\value{
A list of microbes' id for dispersal vanguards and laggards.
}
\description{
This function computes a model-driven testable boundary between dispersal vanguards and laggards
in the Temporary Transition Zone (TTZ). It assigns every microbe within TTZ into either mu-wing 
(i.e., the sub-community of dispersal vanguards) or k-wing (i.e., the sub-community of dispersal laggards).
}
\examples{
# import data
data(hzmicrobe)
param_trio_bay <- calc_mle_trio(mic_bay, n_sample=10, replicates=3)
# classify_ttz is wrapped up in function classify_vag_lag
id_vag_lag <- classify_vag_lag(mic_bay, param_trio_bay)
# print the ids for dispersal vanguards
print(id_vag_lag$vanguards)
# print the ids for dispersal laggards
print(id_vag_lag$laggards)

}
