% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_vag_lag.R
\name{classify_vag_lag}
\alias{classify_vag_lag}
\title{Refine the boundary between sub-communities.}
\usage{
classify_vag_lag(x, mle_trio)
}
\arguments{
\item{x}{The n by m matrix of microbial abundance.}

\item{mle_trio}{Then n by 3 matrix of MLE estimates.}
}
\value{
A list of microbes' id for dispersal vanguards and laggards.
}
\description{
This function computes a model-driven testable boundary between dispersal vanguards and laggards.
It assigns every microbe within the whole community into either mu-wing (i.e., the sub-community of dispersal vanguards) 
or k-wing (i.e., the sub-community of dispersal laggards).
}
\examples{
# import data
data(hzmicrobe)
param_trio_bay <- calc_mle_trio(mic_bay, n_sample=10, replicates=3)
id_vag_lag <- classify_vag_lag(mic_bay, param_trio_bay)
# print the ids for dispersal vanguards
print(id_vag_lag$vanguards)
# print the ids for dispersal laggards
print(id_vag_lag$laggards)

}
