% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{count_to_preabs}
\alias{count_to_preabs}
\title{Diversity Evaluation}
\usage{
count_to_preabs(x)
}
\arguments{
\item{x}{The $n$ by $m$ matrix of microbial abundance.}
}
\description{
A set of functions that assist to calculate diversity indeces including
alpha diversity index
gamma diversity index
beta diversity index
}
