% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_prior_parameters_mle.R
\name{is_dispersion}
\alias{is_dispersion}
\title{MLE Parameters Trio}
\usage{
is_dispersion(v)
}
\arguments{
\item{x}{The $n$ by $m$ matrix of microbial abundance.}
}
\value{
A trio of two estimates of gamma-poisson parameters and the coefficient of zero-inflated model.
}
\description{
This function computes the MLE of prior parameters
in the Gamma-Poisson model (GPM)
}
