% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trio.R
\name{plot_trio}
\alias{plot_trio}
\title{Plot Parameters Trio to Observe the Two-Wing Pattern}
\usage{
plot_trio(param_trio, point_size = 0.8, a = 0.6)
}
\arguments{
\item{param_trio}{The n by 3 matrix of MLE parameter trio estimates.}

\item{point_size}{The numeric number between 0 and 1 for the point size.}

\item{a}{The numeric number between 0 and 1 for the alpha transparency level.}
}
\value{
Nothing of interest.
}
\description{
This function provides graphic display of the MLE parameter trio.
}
\examples{
# import data
data(hzmicrobe)
param_trio_bay <- calc_mle_trio(mic_bay, n_sample=10, replicates=3)
plot_trio(param_trio_bay)


}
