% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vag_lag.R
\name{plot_vag_lag}
\alias{plot_vag_lag}
\title{Visualize the Two-Wing admixed structure}
\usage{
plot_vag_lag(
  param_trio,
  id_category,
  point_size = 0.8,
  a = 0.6,
  zoom_in = FALSE
)
}
\arguments{
\item{param_trio}{The n by 3 matrix of MLE parameter trio estimates.}

\item{point_size}{The numeric number for the point size.}

\item{a}{The numeric number between 0 and 1 for the alpha transparency level.}

\item{zoom_in}{If TRUE, zoom in the majority of points. Default is FALSE.}
}
\value{
Nothing of interest.
}
\description{
This function provides graphic display of the admixed structure of dispersal vanguards and laggards.
}
\examples{
# import data
data(hzmicrobe)
param_trio_bay <- calc_mle_trio(mic_bay, n_sample=10, replicates=3)
id_vag_lag <- classify_vag_lag(mic_bay, param_trio_bay)
plot_vag_lag(param_trio_bay, id_vag_lag)


}
