% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrn_annotate.R
\name{rrn_anno}
\alias{rrn_anno}
\title{Annotate ribosome RNA operons}
\usage{
rrn_anno(tax, rrn)
}
\arguments{
\item{tax}{The $n$ by $3$ matrix of taxa id, family, and genus.}
}
\value{
A vector of mean value of rrn copy number.
}
\description{
This function annotates taxa with rrn count
}
