% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mle_trio.R
\name{calc_mle_trio}
\alias{calc_mle_trio}
\title{MLE Parameters Trio}
\usage{
calc_mle_trio(
  x,
  n_sample = NULL,
  replicates = NULL,
  max_prop_zero = 2/3,
  trace = FALSE
)
}
\arguments{
\item{x}{The n by m matrix of microbial abundance.}

\item{n_sample}{The numeric number of sample size.}

\item{replicates}{The numeric number of replicates per sample site.}

\item{max_prop_zero}{The maximum acceptable value of zero proportions per sample site.}

\item{trace}{If TRUE, print out optimization details. Default is FALSE.}
}
\value{
A trio of two estimates of gamma-poisson parameters and the coefficient of zero-inflated model.
}
\description{
This function computes the MLE of prior parameters
in the Gamma-Poisson model (GPM)
}
\examples{
# import data
data(hzmicrobe)
param_bay <- calc_mle_trio(mic_bay, n_sample=10, replicates=3)


}
